Rewrote MauiImageListTests
