+++ Winforms.sln
@@ -134,8 +134,6 @@ Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "DemoConsole", "src\System.W
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "DesignSurfaceExt", "src\System.Windows.Forms\tests\IntegrationTests\DesignSurface\DesignSurfaceExt\DesignSurfaceExt.csproj", "{E96C74BA-9F74-4289-BF72-45CAD472D3D2}"
 EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "MauiImageListTests", "src\System.Windows.Forms\tests\IntegrationTests\MauiTests\ImageListTests\MauiImageListTests.csproj", "{0B4C8C7D-6157-46D4-AC1E-DF27F96C7AF6}"
-EndProject
 Project("{778DAE3C-4631-46EA-AA77-85C1314464D9}") = "Microsoft.VisualBasic.Forms.Tests", "src\Microsoft.VisualBasic.Forms\tests\UnitTests\Microsoft.VisualBasic.Forms.Tests.vbproj", "{FC75CB54-D8D0-4B41-9A4D-9F862F34A02D}"
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "System.Windows.Forms.Common.TestUtilities", "src\Common\tests\TestUtilities\System.Windows.Forms.Common.TestUtilities.csproj", "{05FD23CE-60AE-44A8-8DD6-1688F04BE385}"
@@ -735,22 +733,6 @@ Global
 		{E96C74BA-9F74-4289-BF72-45CAD472D3D2}.Release|x64.Build.0 = Release|Any CPU
 		{E96C74BA-9F74-4289-BF72-45CAD472D3D2}.Release|x86.ActiveCfg = Release|Any CPU
 		{E96C74BA-9F74-4289-BF72-45CAD472D3D2}.Release|x86.Build.0 = Release|Any CPU
-		{0B4C8C7D-6157-46D4-AC1E-DF27F96C7AF6}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{0B4C8C7D-6157-46D4-AC1E-DF27F96C7AF6}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{0B4C8C7D-6157-46D4-AC1E-DF27F96C7AF6}.Debug|arm64.ActiveCfg = Debug|Any CPU
-		{0B4C8C7D-6157-46D4-AC1E-DF27F96C7AF6}.Debug|arm64.Build.0 = Debug|Any CPU
-		{0B4C8C7D-6157-46D4-AC1E-DF27F96C7AF6}.Debug|x64.ActiveCfg = Debug|Any CPU
-		{0B4C8C7D-6157-46D4-AC1E-DF27F96C7AF6}.Debug|x64.Build.0 = Debug|Any CPU
-		{0B4C8C7D-6157-46D4-AC1E-DF27F96C7AF6}.Debug|x86.ActiveCfg = Debug|Any CPU
-		{0B4C8C7D-6157-46D4-AC1E-DF27F96C7AF6}.Debug|x86.Build.0 = Debug|Any CPU
-		{0B4C8C7D-6157-46D4-AC1E-DF27F96C7AF6}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{0B4C8C7D-6157-46D4-AC1E-DF27F96C7AF6}.Release|Any CPU.Build.0 = Release|Any CPU
-		{0B4C8C7D-6157-46D4-AC1E-DF27F96C7AF6}.Release|arm64.ActiveCfg = Release|Any CPU
-		{0B4C8C7D-6157-46D4-AC1E-DF27F96C7AF6}.Release|arm64.Build.0 = Release|Any CPU
-		{0B4C8C7D-6157-46D4-AC1E-DF27F96C7AF6}.Release|x64.ActiveCfg = Release|Any CPU
-		{0B4C8C7D-6157-46D4-AC1E-DF27F96C7AF6}.Release|x64.Build.0 = Release|Any CPU
-		{0B4C8C7D-6157-46D4-AC1E-DF27F96C7AF6}.Release|x86.ActiveCfg = Release|Any CPU
-		{0B4C8C7D-6157-46D4-AC1E-DF27F96C7AF6}.Release|x86.Build.0 = Release|Any CPU
 		{FC75CB54-D8D0-4B41-9A4D-9F862F34A02D}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{FC75CB54-D8D0-4B41-9A4D-9F862F34A02D}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{FC75CB54-D8D0-4B41-9A4D-9F862F34A02D}.Debug|arm64.ActiveCfg = Debug|Any CPU
@@ -959,7 +941,6 @@ Global
 		{43E46506-7DF8-4E7A-A579-996CA43041EB} = {680FB14C-7B0C-4D63-9F1A-18ACCDB0F52A}
 		{93310A19-DDCA-4BCD-AEDE-5C5D788DAFB4} = {43E46506-7DF8-4E7A-A579-996CA43041EB}
 		{E96C74BA-9F74-4289-BF72-45CAD472D3D2} = {43E46506-7DF8-4E7A-A579-996CA43041EB}
-		{0B4C8C7D-6157-46D4-AC1E-DF27F96C7AF6} = {8F20A905-BD37-4D80-B8DF-FA45276FC23F}
 		{FC75CB54-D8D0-4B41-9A4D-9F862F34A02D} = {583F1292-AE8D-4511-B8D8-A81FE4642DDC}
 		{05FD23CE-60AE-44A8-8DD6-1688F04BE385} = {DF68A171-D27B-4E6A-8A7E-63A651622355}
 		{5D8C7F71-8B58-48B4-8A54-11E5AB8B02CA} = {8F20A905-BD37-4D80-B8DF-FA45276FC23F}
